+++ packages/core/src/createEventChain.ts
@@ -1,120 +1,119 @@
 import * as S from "./types"
 import { createDraft, finishDraft, Draft, original } from "immer"
 
-type Core<D> = {
-  data: D
-  result: any
-}
-
-type Callback<D> = (info: Outcome<D>) => void
-
-type Outcome<D> = {
-  data: D
-  shouldBreak: boolean
-  shouldNotify: boolean
-  pendingTransition?: string
-  pendingSend?: S.Event
-}
-
-type Options<D> = {
-  data: D
-  result: any
-  payload: any
-  handler: S.EventHandler<D>
-  onAsyncUpdate: Callback<D>
-  onRefreshDataAfterWait: () => D
-}
-
-export function createEventChain<D>(options: Options<D>) {
-  let { onAsyncUpdate, onRefreshDataAfterWait } = options
+export function createEventChain<D>(options: S.EventChainOptions<D>) {
+  let { state, onDelayedOutcome, getFreshDataAfterWait } = options
   let handlers = [...options.handler]
   const { payload } = options
 
   let waiting = false
 
-  let finalOutcome: Outcome<D> = {
+  let core: S.EventChainCore<D> = {
     data: options.data,
+    result: options.result,
+  }
+
+  let finalOutcome: S.EventChainOutcome<D> = {
+    ...core,
     shouldBreak: false,
     shouldNotify: false,
     pendingSend: undefined,
     pendingTransition: undefined,
   }
 
-  let core = {
-    data: options.data,
-    result: options.result,
-  }
+  let draftCore: Draft<S.EventChainCore<D>> = createDraft(core)
 
-  let draftCore: Draft<Core<D>> = createDraft(core)
-  let orig = original(draftCore) as Draft<Core<D>>
+  let orig = original(draftCore) as Draft<S.EventChainCore<D>>
+
   let tResult = orig?.result
 
-  function complete(draft: Draft<Core<D>>) {
-    core = finishDraft(draft) as Core<D>
+  function complete(draft: Draft<S.EventChainCore<D>>) {
+    core = finishDraft(draft) as S.EventChainCore<D>
+    finalOutcome.result = core.result
     finalOutcome.data = core.data
   }
 
   function refresh() {
     draftCore = createDraft(core)
-    orig = original(draftCore) as Draft<Core<D>>
+    orig = original(draftCore) as Draft<S.EventChainCore<D>>
     tResult = orig?.result
   }
 
   function processEventHandler(
     eventHandler: S.EventHandler<D>,
-    draft: Draft<Core<D>>
-  ) {
+    draft: Draft<S.EventChainCore<D>>
+  ): { shouldBreakDueToWait: boolean } {
     if (finalOutcome.shouldBreak) {
-      return
+      return { shouldBreakDueToWait: false }
     }
 
     const nextHandlerObject = eventHandler.shift()
 
     if (nextHandlerObject === undefined) {
-      return
+      return { shouldBreakDueToWait: false }
     } else if (nextHandlerObject.wait !== undefined) {
       // Calculate wait time from finalOutcome draft
       const waitTime =
         nextHandlerObject.wait((draft.data as D) as D, payload, draft.result) *
         1000
 
-      // // Notify, if necessary
+      // Notify, if necessary
       if (waiting && finalOutcome.shouldNotify) {
         complete(draftCore)
-        onAsyncUpdate(finalOutcome)
-        refresh()
+        onDelayedOutcome(finalOutcome)
       }
 
       waiting = true
 
-      setTimeout(() => {
-        core.data = onRefreshDataAfterWait() // After the timeout, refresh data
+      // TODO: Does timeouts need to be an array?
+      state.times.timeouts[0] = setTimeout(() => {
+        core.data = getFreshDataAfterWait() // After the timeout, refresh data
         core.result = undefined // Results can't be carried across!
 
         refresh()
 
-        processHandlerObject(nextHandlerObject, draftCore)
-        processEventHandler(handlers, draftCore)
+        const { shouldBreak } = processHandlerObject(
+          nextHandlerObject,
+          draftCore
+        )
+
+        if (!shouldBreak) {
+          const { shouldBreakDueToWait } = processEventHandler(
+            handlers,
+            draftCore
+          )
+
+          if (shouldBreakDueToWait) {
+            // If the event handler produced a wait, then it
+            // will have also refreshed the core and notified
+            // subscribers, if necessary.
+            return
+          }
+        }
 
         complete(draftCore)
-        onAsyncUpdate(finalOutcome)
+        onDelayedOutcome(finalOutcome)
       }, waitTime)
+
+      return { shouldBreakDueToWait: true }
 
       // Stop this chain
     } else {
       // Continue with chain
-      const shouldContinue = processHandlerObject(nextHandlerObject, draft)
+      const { shouldBreak } = processHandlerObject(nextHandlerObject, draft)
 
-      if (shouldContinue) {
-        processEventHandler(eventHandler, draft)
+      if (!shouldBreak) {
+        return processEventHandler(eventHandler, draft)
       }
+
+      return { shouldBreakDueToWait: false }
     }
   }
 
   function processHandlerObject(
     handler: S.EventHandlerObject<D>,
-    draft: Draft<Core<D>>
-  ) {
+    draft: Draft<S.EventChainCore<D>>
+  ): { shouldBreak: boolean } {
     let passedConditions = true
 
     // Compute a result using original data and draft result
@@ -187,7 +186,7 @@ export function createEventChain<D>(options: Options<D>) {
 
         finalOutcome.shouldBreak = true
         finalOutcome.shouldNotify = true
-        return false
+        return { shouldBreak: true }
       }
 
       // Secret Transitions
@@ -207,7 +206,7 @@ export function createEventChain<D>(options: Options<D>) {
       // Break
       if (handler.break !== undefined) {
         if (handler.break(draft.data as D, payload, tResult)) {
-          return false
+          return { shouldBreak: true }
         }
       }
     } else {
@@ -217,7 +216,7 @@ export function createEventChain<D>(options: Options<D>) {
       }
     }
 
-    return true
+    return { shouldBreak: false }
   }
 
   processEventHandler(handlers, draftCore)