+++ packages/tutorials/src/tutorials/dogs/starter/index.js
@@ -1,0 +1,13 @@
+import React from "react"
+import { Dogs as D } from "components"
+import { delay } from "utils"
+
+export default function () {
+  return (
+    <D.Layout>
+      <D.Image image={""} />
+      <D.Button isLoading={false}>Fetch</D.Button>
+      <D.Button disabled={true}>Cancel</D.Button>
+    </D.Layout>
+  )
+}